backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
google-generativeai>=0.8.0
elevenlabs>=1.15.0
pydantic>=2.10.0
python-dotenv>=1.0.0
